requirements.txt
numpy>=1.21.0
pillow>=9.0.0
scikit-image>=0.19.0
trimesh>=3.9.0
scipy>=1.7.0
tqdm>=4.62.0
